improve(sdk): refactor tables to be more consistent
