+++ packages/sdk/src/libs/stores/sorted-js-map/README.md
@@ -2,7 +2,7 @@
 
 This wraps a JS map, but also includes a sorted array which allows you to query data between keys. Useful
 for time series data or anything that needs to be ordered. Uses lodash binary searches under the hood
-so queries and updates should be order Nlog(N).
+so queries and updates should be O(log(N)).
 
 ## Usage
