+++ packages/sdk/src/libs/tables/base/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./base";
+export { default } from "./table";
