+++ packages/sdk/src/libs/tables/blocks/index.ts
@@ -1,9 +1,2 @@
-export * from "./js-map";
-
-export type Data = {
-  id?: number;
-  number: number;
-  hash: string;
-  // block time is in seconds
-  timestamp: number;
-};
+export { Table as JsMap } from "./js-map";
+export * from "./utils";
