+++ packages/sdk/src/libs/tables/blocks/utils.ts
@@ -1,0 +1,11 @@
+export function makeId(data: Data) {
+  return data.number;
+}
+
+export type Data = {
+  id?: number;
+  number: number;
+  hash: string;
+  // block time is in seconds
+  timestamp: number;
+};
