+++ packages/sdk/src/libs/tables/emps/index.ts
@@ -1,26 +1,2 @@
-export * from "./js-map";
-
-export type Data = {
-  id?: string;
-  name?: string | null;
-  address: string;
-  priceIdentifier?: string | null;
-  expirationTimestamp?: string | null;
-  withdrawLiveness?: string | null;
-  tokenCurrency?: string | null;
-  collateralCurrency?: string | null;
-  collateralRequirement?: string | null;
-  disputeBondPercentage?: string | null;
-  sponsorDisputeRewardPercentage?: string | null;
-  disputerDisputeRewardPercentage?: string | null;
-  cumulativeFeeMultiplier?: string | null;
-  tokenDecimals?: number | null;
-  collateralDecimals?: number | null;
-  totalTokensOutstanding?: string | null;
-  totalPositionCollateral?: string | null;
-  minSponsorTokens?: string | null;
-  expiryPrice?: string | null;
-  sponsors?: string[] | null;
-  gcr?: string | null;
-  expired?: boolean | null;
-};
+export { Table as JsMap } from "./js-map";
+export * from "./utils";
