+++ packages/sdk/src/libs/tables/emps/js-map.ts
@@ -1,11 +1,8 @@
-import type { Data } from ".";
-import { JsMap as GenericJsMap } from "../generic";
+import { Data, makeId } from ".";
+import { JsMap } from "../generic";
 
-export const JsMap = (type = "Emp") => {
-  function makeId(data: Data) {
-    return data.address;
-  }
-  const table = GenericJsMap<string, Data>(type, makeId);
+export const Table = (type = "Emp") => {
+  const table = JsMap<string, Data>(type, makeId);
 
   async function addSponsors(id: string, sponsors: string[]) {
     const data = await table.get(id);
@@ -20,4 +17,4 @@ export const JsMap = (type = "Emp") => {
   };
 };
 // want to export the type as the same name
-export type JsMap = ReturnType<typeof JsMap>;
+export type Table = ReturnType<typeof Table>;
