+++ packages/sdk/src/libs/tables/emps/utils.ts
@@ -1,0 +1,28 @@
+export function makeId(data: Data) {
+  return data.address;
+}
+
+export type Data = {
+  id?: string;
+  name?: string | null;
+  address: string;
+  priceIdentifier?: string | null;
+  expirationTimestamp?: string | null;
+  withdrawLiveness?: string | null;
+  tokenCurrency?: string | null;
+  collateralCurrency?: string | null;
+  collateralRequirement?: string | null;
+  disputeBondPercentage?: string | null;
+  sponsorDisputeRewardPercentage?: string | null;
+  disputerDisputeRewardPercentage?: string | null;
+  cumulativeFeeMultiplier?: string | null;
+  tokenDecimals?: number | null;
+  collateralDecimals?: number | null;
+  totalTokensOutstanding?: string | null;
+  totalPositionCollateral?: string | null;
+  minSponsorTokens?: string | null;
+  expiryPrice?: string | null;
+  sponsors?: string[] | null;
+  gcr?: string | null;
+  expired?: boolean | null;
+};
