+++ packages/sdk/src/libs/tables/generic/index.ts
@@ -1,1 +1,2 @@
 export { default as JsMap } from "./js-map";
+export { default as SortedJsMap } from "./sorted-js-map";
