+++ packages/sdk/src/libs/tables/generic/sorted-js-map.ts
@@ -1,0 +1,8 @@
+import Base from "../base";
+import Store from "../../stores/sorted-js-map";
+
+export default function <I, D>(type: string, makeId: (d: D) => I) {
+  const store = Store<I, D>();
+  const table = Base<I, D>({ type, makeId }, store);
+  return table;
+}
