+++ packages/sdk/src/libs/utils.ts
@@ -1,5 +1,5 @@
 import assert from "assert";
-import { BigNumber, utils } from "ethers";
+import { BigNumber } from "ethers";
 
 export type BigNumberish = number | string | BigNumber;
 // check if a value is not null or undefined, useful for numbers which could be 0.
